[Bugfix] instant ammotype changing fix
When you try to reload weapon but count of ammos in your inventory is not enought for full magazine, ammo's icon automatically changes to an icon of another ammotype.
Also the bug could lead to another kind of behavior. Take into your hands a weapon with not full mag and without ammo of current ammo type in the inventory. Start reloading it and (without waiting for finish of reloading) drop the weapon. Then take it back to the hands. Changing of ammotype would be broken.